blocked user auto logout

+++ frontend/src/pages/Homepage.jsx
@@ -6,6 +6,7 @@ import ButtonBlock from "../components/blockButton";
 import updateStatus from "../api/users/updateStatus";
 import deleteUsers from "../api/users/deleteUsers";
 import {useNavigate} from "react-router-dom";
+import {getUser} from "../api/users/getCurrentUser";
 
 const Homepage = () => {
 
@@ -13,6 +14,8 @@ const Homepage = () => {
     const [users, setUsers] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
     const [userIds, setUserIds] = useState([]);
+
+    const [currentUserStatus, setCurrentUserStatus] = useState([]);
 
     const navigate = useNavigate();
 
@@ -25,19 +28,41 @@ const Homepage = () => {
 
     }
 
+    async function getCurrentUser() {
+        const currentUserData = await getUser(token);
+        console.log(currentUserData.status);
+        setCurrentUserStatus(currentUserData.status);
+
+    }
+
     useEffect(() => {
         getUsers();
+        getCurrentUser();
     }, []);
 
     async function updateUserStatus(status) {
-        await updateStatus(userIds, status, token, successReload);
-        getUsers();
+        if (currentUserStatus === "blocked") {
+            alert("You are blocked");
+            setTimeout(() => {
+                navigate("/signUp")
+            }, 2000)
+        } else {
+            await updateStatus(userIds, status, token, successReload);
+            getUsers();
+        }
     }
 
     async function handleDeleteUsers() {
-        await deleteUsers(userIds, successReload);
-        console.log(userIds);
-        getUsers();
+        if (currentUserStatus === "blocked") {
+            alert("You are blocked");
+            setTimeout(() => {
+                navigate("/signUp")
+            }, 2000)
+        } else {
+            await deleteUsers(userIds, successReload);
+            console.log(userIds);
+            getUsers();
+        }
     }
 
     const handleSelectedChange = (newSelected) => {
